Make a Data Set of GSOD Data Suitable for Interpolation

This function can be wrapped in an [base::lapply()] function to retrieve and save multiple years of GSOD data for interpolation, though it may be used to retrieve GSOD data and interpolate files on-the-fly. See vignette("GRID") for more details and examples.

make_GSOD_set(years = NULL, dsn = NULL)

Arguments

years
A numeric vector of years of GSOD data to get for interpolation. Defaults to current year.

dsn
Optional. A filepath where resulting CSV files are to be saved on local disk. If unspecified a tidy data frame is returned in the R session.

Value

A [base::list()] of [base::data.frame] objects containing GSOD data suitable for interpolation using [interpolate_gsod()]

Details

This function will get GSOD data using [GSODR::get_GSOD()] and save a CSV file containing only the following fields to use in interpolating a global surface between 60 and -60 degrees latitude.

STNID

A unique station id number that can be used to identify the station and link with GSODR data for station metadata.

LON

Longitude in decimal degrees.

LAT

Latitude in decimal degrees.

ELEV_M_SRTM_90m

Elevation in metres corrected for possible errors, derived from the CGIAR-CSI SRTM 90m database (Jarvis et al. 2008).

YEAR

The year (YYYY).

YDAY

Sequential day of year.

TEMP

Mean daily temperature in degrees C to tenths. Missing = NA.

MAX

Maximum temperature reported during the day in Celsius to tenths. Missing = NA;

MIN

Minimum temperature reported during the day in Celsius to tenths. Missing = NA;

RH

Mean daily relative humidity. Missing = NA.

References

Jarvis, A., Reuter, H. I., Nelson, A., Guevara, E. (2008) Hole-filled SRTM for the globe Version 4, available from the CGIAR-CSI SRTM 90m Database (http://srtm.csi.cgiar.org)

Examples

# NOT RUN {
# Get one year of GSOD data
gsod_1998 <- make_GSOD_set(years = 1998)

# Get multiple years of GSOD data and save to disk
years <- as.list(seq(from = 1983, to = 2017, by = 1))
lapply(X = years, FUN = make_GSOD_set, dsn = "~/Data/GSOD")
# }